- type conversion fixes

+++ src/rendersystems/OpenGL/GLES/OpenGLES_RenderSystem.cpp
@@ -691,7 +691,7 @@ namespace april
 			return;
 		}
 		glBindFramebuffer(GL_FRAMEBUFFER, ((OpenGLES_Texture*)destination)->framebufferId);
-		this->_intermediateState->viewport.setSize((float)source->getWidth(), (float)source->getHeight());
+		this->_intermediateState->viewport.setSize(source->getWidth(), source->getHeight());
 		this->_intermediateState->projectionMatrix.setOrthoProjection(
 			grectf(1.0f - 2.0f * this->pixelOffset / source->getWidth(), 1.0f - 2.0f * this->pixelOffset / source->getHeight(), 2.0f, 2.0f));
 		this->_intermediateState->texture = source;
